PIM-5384: Extract what's generic from NavigationContext
